Change `replace` arguments to functions
Because QA Wolf at some point added the new `execa` syntax (`$\``), but
dollar signs have a special meaning when used inside of the string that
is passed to the right hand side of `replace` because... reasons. But if
the string is returned from a function it's just a plain string. Don't
look at me. I don't make the rules.